Summarizer/requirements.txt
flask
flask-cors
pdfplumber
python-docx
transformers
torch
sentence-transformers
google-generativeai
numpy
rouge-score
nltk 